separate files for entrez,serratus api calls
